Now the player can shoot only 5 projectile + Enemy will now be teleported
After the player shoot 5 projectiles if none one of them hit the enemy, the player will lose the game. Each time a projectile is destroyed the enemy will be teleported in a cell within the fog of war if there's one, otherwise it will be teleported in a random position.